backend/requirements.txt
# LangGraph Text Analysis Backend - Requirements
# For deployment on Render.com

# Core LangGraph & LangChain
langgraph>=0.2.62
langchain>=0.3.18
langchain-ollama>=0.2.0
langchain-chroma>=0.1.0
chromadb>=0.5.0
langchain-core>=0.3.29
typing-extensions>=4.12.0

# FastAPI & Web Server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.6.0

# HTTP & CORS
python-multipart>=0.0.9
httpx>=0.27.0

# Environment & Configuration
python-dotenv>=1.0.0

# Optional: For better performance
orjson>=3.10.0
